+++ cognite/experimental/data_classes/functions.py
@@ -20,6 +20,7 @@ class Function(CogniteResource):
         created_time (int): Created time in UNIX.
         api_key (str): Api key attached to the function.
         secrets (Dict[str, str]): Secrets attached to the function ((key, value) pairs).
+        env_vars (Dict[str, str]): User specified environment variables on the function ((key, value) pairs).
         cpu (Number): Number of CPU cores per function. Defaults to 0.25. Allowed values are in the range [0.1, 0.6].
         memory (Number): Memory per function measured in GB. Defaults to 1. Allowed values are in the range [0.1, 2.5].
         error(Dict[str, str]): Dictionary with keys "message" and "trace", which is populated if deployment fails.
@@ -39,6 +40,7 @@ class Function(CogniteResource):
         created_time: int = None,
         api_key: str = None,
         secrets: Dict = None,
+        env_vars: Dict = None,
         cpu: Number = None,
         memory: Number = None,
         error: Dict = None,
@@ -55,6 +57,7 @@ class Function(CogniteResource):
         self.created_time = created_time
         self.api_key = api_key
         self.secrets = secrets
+        self.env_vars = env_vars
         self.cpu = cpu
         self.memory = memory
         self.error = error