+++ cognite/experimental/data_classes/__init__.py
@@ -4,5 +4,5 @@ from cognite.experimental.data_classes.contextualization import *
 from cognite.experimental.data_classes.functions import *
 from cognite.experimental.data_classes.integrationruns import *
 from cognite.experimental.data_classes.integrations import *
-from cognite.experimental.data_classes.relationships import *
+from cognite.experimental.data_classes.templates import *
 from cognite.experimental.data_classes.types import *