+++ CHANGELOG.md
@@ -12,6 +12,14 @@ Changes are grouped as follows
 - `Fixed` for any bug fixes.
 - `Security` in case of vulnerabilities.
 
+## [0.42.0] - 21-02-01
+
+### Added 
+- Template functionality.
+
+### Removed
+- Playground relationships.
+
 ## [0.41.0] - 2020-01-19
 
 ### Added